Modbus2DBs/Bilancia_Serial_2_Influx/requirements.txt
azure-iot-device==2.4.0
pymodbus
numpy
aenum
pyparsing
psycopg2-binary
sqlalchemy
asyncio
datetime
influxdb
six
pyparsing
pandas
pyserial
